webkit/LayoutTests/platform/blackberry/fast/events/pass-escape-key-to-non-editable-node-expected.txt

This tests that the escape keyboard key, triggered by the BlackBerry Escape (back) button, is passed to a non-text input/content-editable node. This test PASSED if you see PASS below. Otherwise, it FAILED.

PASS keyCode is 27.

